Move MaxText related util functions to maxtext_utils.py.
