+++ MaxText/scratch_code/generate_grpo_golden_logits.py
@@ -31,6 +31,7 @@ import jax.numpy as jnp
 import numpy as np
 
 from MaxText import pyconfig
+from MaxText import maxtext_utils
 from MaxText import max_utils
 from jax.sharding import Mesh
 from flax import linen as nn
@@ -77,16 +78,16 @@ class GRPOTest(unittest.TestCase):
         per_device_batch_size=self.cfg_no_ckpt_loading.per_device_batch_size * self.cfg_no_ckpt_loading.num_generations,
     )
     self.rng = jax.random.key(self.cfg.init_weights_seed)
-    devices_array = max_utils.create_device_mesh(self.cfg)
+    devices_array = maxtext_utils.create_device_mesh(self.cfg)
     mesh = Mesh(devices_array, self.cfg.mesh_axes)
     # With checkpoint
     self.model = models.Transformer(config=self.cfg, mesh=mesh, quant=None)
-    self.state, state_mesh_annotations = max_utils.setup_decode_state(self.model, self.cfg, self.rng, mesh, None)
+    self.state, state_mesh_annotations = maxtext_utils.setup_decode_state(self.model, self.cfg, self.rng, mesh, None)
     self.state_mesh_shardings = nn.logical_to_mesh_sharding(state_mesh_annotations, mesh, self.cfg.logical_axis_rules)
     self.data_sharding = jax.NamedSharding(mesh, jax.sharding.PartitionSpec(None))
     # Without checkpoint
     self.model_no_ckpt_loading = models.Transformer(config=self.cfg_no_ckpt_loading, mesh=mesh, quant=None)
-    self.state_no_ckpt_loading, _ = max_utils.setup_decode_state(
+    self.state_no_ckpt_loading, _ = maxtext_utils.setup_decode_state(
         self.model_no_ckpt_loading, self.cfg_no_ckpt_loading, self.rng, mesh, None
     )
