+++ MaxText/tests/forward_pass_logit_checker.py
@@ -50,6 +50,7 @@ import numpy as np
 from MaxText import pyconfig
 import jsonlines
 from MaxText import max_utils
+from MaxText import maxtext_utils
 from MaxText.layers import models
 from MaxText.layers import quantizations
 
@@ -88,11 +89,11 @@ def main(config, test_args):  # pylint: disable=W0621
   else:  # Initialize MaxText model
     init_rng = jax.random.PRNGKey(config.init_weights_seed)
     init_rng, rng1 = jax.random.split(init_rng)
-    devices_array = max_utils.create_device_mesh(config)
+    devices_array = maxtext_utils.create_device_mesh(config)
     mesh = jax.sharding.Mesh(devices_array, config.mesh_axes)
     quant = quantizations.configure_quantization(config)
     model = models.Transformer(config, mesh=mesh, quant=quant)
-    state, _ = max_utils.setup_decode_state(model, config, rng1, mesh, None)
+    state, _ = maxtext_utils.setup_decode_state(model, config, rng1, mesh, None)
 
   if test_args.golden_logits_path == "":
     input_golden_data_path = os.path.join(PKG_DIR, "test_assets", f"golden_data_{config.model_name}.jsonl")
