+++ MaxText/tests/grpo_trainer_correctness_test.py
@@ -34,7 +34,7 @@ from flax import linen as nn
 from MaxText.globals import PKG_DIR
 
 
-from MaxText import max_utils
+from MaxText import maxtext_utils
 from MaxText import pyconfig
 from MaxText.layers import models
 from MaxText.layers import quantizations
@@ -56,10 +56,10 @@ def get_golden_data(config):
 def setup_maxtext_model(config):
   init_rng = jax.random.PRNGKey(config.init_weights_seed)
   quant = quantizations.configure_quantization(config)
-  devices_array = max_utils.create_device_mesh(config)
+  devices_array = maxtext_utils.create_device_mesh(config)
   mesh = Mesh(devices_array, config.mesh_axes)
   maxtext_model = models.Transformer(config=config, mesh=mesh, quant=quant)
-  state, state_mesh_annotations = max_utils.setup_decode_state(maxtext_model, config, init_rng, mesh, None)
+  state, state_mesh_annotations = maxtext_utils.setup_decode_state(maxtext_model, config, init_rng, mesh, None)
   state_mesh_shardings = nn.logical_to_mesh_sharding(state_mesh_annotations, mesh, config.logical_axis_rules)
   data_sharding = jax.NamedSharding(mesh, jax.sharding.PartitionSpec(None))
   reference_params = jax.tree.map(jnp.copy, state.params["params"])
