+++ MaxText/tests/integration_tests/grpo_correctness.py
@@ -21,7 +21,7 @@ import jax.numpy as jnp
 import numpy as np
 
 from MaxText import pyconfig
-from MaxText import max_utils
+from MaxText import maxtext_utils
 from jax.sharding import Mesh
 import flax.linen as nn
 from typing import Tuple
@@ -64,10 +64,10 @@ class GRPOTest(unittest.TestCase):
         logits_dot_in_fp32=True,
     )
     self.rng = jax.random.PRNGKey(42)
-    devices_array = max_utils.create_device_mesh(self.cfg)
+    devices_array = maxtext_utils.create_device_mesh(self.cfg)
     mesh = Mesh(devices_array, self.cfg.mesh_axes)
     self.model = models.Transformer(config=self.cfg, mesh=mesh, quant=None)
-    self.state, _ = max_utils.setup_decode_state(self.model, self.cfg, self.rng, mesh, None)
+    self.state, _ = maxtext_utils.setup_decode_state(self.model, self.cfg, self.rng, mesh, None)
     self.tokenizer_model = transformers.AutoTokenizer.from_pretrained(
         "meta-llama/Llama-3.1-8B",
         add_bos_token=False,
